Move ValueConverter to structured
